Got the best way to represent decimal datagrid columns

+++ GestionePosizioni/PosizioniRoverfrutta/CustomControls/DataGridColumns/DataGridDecimalColumn.cs
@@ -17,7 +17,7 @@ namespace PosizioniRoverfrutta.CustomControls.DataGridColumns
         {
             try
             {
-                Convert.ToDecimal(e.Text);
+                Convert.ToDouble(e.Text);
             }
             catch
             {
